Rename module for narrower scope
