+++ server/npg/porch/endpoints/pipelines.py
@@ -22,7 +22,7 @@ from fastapi import APIRouter, HTTPException, Depends
 from typing import List
 
 from npg.porch.models.pipeline import Pipeline
-from npg.porchdb.db import get_DbAccessor
+from npg.porchdb.connection import get_DbAccessor
 
 router = APIRouter(
     prefix="/pipelines",
