+++ server/npg/porch/endpoints/tasks.py
@@ -25,7 +25,7 @@ from typing import List
 from npg.porch.models.pipeline import Pipeline
 from npg.porch.models.task import Task
 
-from npg.porchdb.db import get_DbAccessor
+from npg.porchdb.connection import get_DbAccessor
 
 router = APIRouter(
     prefix="/tasks",
