+++ server/tests/fixtures/orm_session.py
@@ -5,7 +5,7 @@ import sqlalchemy
 import sqlalchemy.orm
 
 from npg.porchdb.models import Base
-from npg.porchdb.db import session_factory, deploy_schema, close_engine
+from npg.porchdb.connection import session_factory, deploy_schema, close_engine
 
 
 @pytest.fixture
